Move MsgParts Display trait test
